requirements.txt
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-WTF==1.2.1
html5lib==1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.2
mysql-connector-python==8.2.0
numpy==1.24.3
pandas==2.0.1
protobuf==4.21.12
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
tzdata==2023.3
urllib3==2.0.2
webencodings==0.5.1
Werkzeug==2.3.4
WTForms==3.1.2
